Remove MyLinkedHashMap and move functionality into InMemoryHistoryManager
